+++ packages/utils/src/index.ts
@@ -1,7 +1,7 @@
 export { callAll } from './utils/CallAll';
 export { throttle } from './utils/throttle';
 export * as dateFormatters from './utils/dates/dateFormatters';
-export * as timeAgo from './utils/time/timeAgo';
+export * as timeAgo from './utils/dates/timeAgo';
 export { mapObjIndexed } from './utils/objects/mapObjIndexed';
 export { generateRange } from './utils/pagination';
 export * from './utils/objects';
