+++ packages/utils/src/utils/dates/index.ts
@@ -1,1 +1,2 @@
 export * as dateFormatters from './dateFormatters';
+export * as timeAgo from './timeAgo';
