--- packages/utils/src/utils/time/index.ts
@@ -1,1 +1,0 @@
-export * as timeAgo from './timeAgo';
