+++ packages/widgets/src/components/NotificationCard/index.ts
@@ -1,2 +1,2 @@
-export { default as NotificationCard } from './NotificationCard';
+export { NotificationCard } from './NotificationCard';
 export * from './types';
