+++ packages/widgets/src/components/NotificationCard/styles.tsx
